pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "baton-orchestrator"
version = "0.1.0"
description = "Cloud-agnostic circuit orchestration -- pre-wired topologies with smart adapters, mock collapse, and self-healing."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
authors = [
    {name = "J. Andrew McEntire", email = "jandrewmcentire@gmail.com"},
]
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
gcp = [
    "google-cloud-run>=0.10",
    "kubernetes>=28.0",
]
aws = [
    "boto3>=1.30",
]
otel = [
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20",
]
mcp = [
    "mcp[cli]>=1.26.0",
]
certs = [
    "cryptography>=41.0",
]

[project.scripts]
baton = "baton.cli:main"
baton-mcp = "baton.mcp_server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/baton"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-x -q"
